Rename SVG2PDF file
