+++ IPython/nbconvert/transformers/__init__.py
@@ -1,7 +1,7 @@
 # Class base Transformers
 from .base import ConfigurableTransformer
 from .convertfigures import ConvertFiguresTransformer
-from .svg2pdf import Svg2PdfTransformer
+from .SVG2PDF import Svg2PdfTransformer
 from .extractfigure import ExtractFigureTransformer
 from .revealhelp import RevealHelpTransformer
 from .latex import LatexTransformer
